requirements.txt
langchain
langchain-community

ebooklib 
beautifulsoup4 
llama-index 
openai 
python-dotenv